requirements.txt
fastapi
python-dotenv
argon2-cffi
tortoise-orm[accel,asyncpg]
pydantic
pyjwt
uvicorn
